% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeShinyFiles.R
\name{makeShinyFiles}
\alias{makeShinyFiles}
\title{Generate data files required for shiny app}
\usage{
makeShinyFiles(
  obj,
  scConf,
  gex.assay = NA,
  gex.slot = c("data", "scale.data", "counts"),
  gene.mapping = FALSE,
  shiny.prefix = "sc1",
  shiny.dir = "shinyApp/",
  default.gene1 = NA,
  default.gene2 = NA,
  default.multigene = NA,
  default.dimred = NA,
  chunk.size = 500,
  markers.all = FALSE,
  markers.top20 = FALSE,
  de.genes = FALSE,
  gene.ranks = FALSE,
  volc.plot = FALSE,
  gene.ont = FALSE,
  pval.cutoff = 0.5,
  num.genes = 400
)
}
\arguments{
\item{obj}{input single-cell object for Seurat (v3+) / SingleCellExperiment 
data or input file path for h5ad / loom files}

\item{scConf}{shinycell config data.table}

\item{gex.assay}{assay in single-cell data object to use for plotting 
gene expression, which must match one of the following:
\itemize{
  \item{Seurat objects}: "RNA" or "integrated" assay, 
    default is "RNA"
  \item{SCE objects}: "logcounts" or "normcounts" or "counts", 
    default is "logcounts"
  \item{h5ad files}: "X" or any assay in "layers",
    default is "X"
  \item{loom files}: "matrix" or any assay in "layers",
    default is "matrix"
}}

\item{gex.slot}{slot in single-cell assay to plot. This is only used 
for Seurat objects (v3+). Default is to use the "data" slot}

\item{gene.mapping}{specifies whether to convert human / mouse Ensembl gene 
IDs (e.g. ENSG000xxx / ENSMUSG000xxx) into "user-friendly" gene symbols. 
Set this to \code{TRUE} if you are using Ensembl gene IDs. Default is 
\code{FALSE} which is not to perform any conversion. Alternatively, users 
can supply a named vector where \code{names(gene.mapping)} correspond 
to the actual gene identifiers in the gene expression matrix and 
\code{gene.mapping} correspond to new identifiers to map to}

\item{shiny.prefix}{specify file prefix}

\item{shiny.dir}{specify directory to create the shiny app in}

\item{default.gene1}{specify primary default gene to show}

\item{default.gene2}{specify secondary default gene to show}

\item{default.multigene}{character vector specifying default genes to 
show in bubbleplot / heatmap}

\item{default.dimred}{character vector specifying the two default dimension 
reductions. Default is to use UMAP if not TSNE embeddings}

\item{chunk.size}{number of genes written to h5file at any one time. Lower 
this number to reduce memory consumption. Should not be less than 10}

\item{markers.all}{boolean flag as to whether to create the 
"Cluster Markers, All" tab and prepare associated Seurat data}

\item{markers.top20}{boolean flag as to whether to create the
"Cluster Markers, Top 20" tab and prepare associated Seurat data}

\item{de.genes}{boolean flag as to whether to create the "Diff. Exp. Genes"
tab and prepre associated Seurat data}

\item{gene.ranks}{boolean flag as to whether to create the "Gene Signature"
tab and prepare the associated Seurat data}

\item{volc.plot}{boolean flag as to whether to create the 
"Diff. Gene Exp., Volcano" tab and prepare associated Seurat data}

\item{gene.ont}{boolean flag as to whether to create the "ToppGene Ontology"
tab and prepate the associated Seurat data}

\item{pval.cutoff}{upper limit of pval to filter cluster gene expression by (pvals
greater than this are filtered out)}

\item{num.genes}{max number of most expressed genes to include in ToppGene query}
}
\value{
data files required for shiny app
}
\description{
Generate data files required for shiny app. Five files will be generated, 
namely (i) the shinycell config \code{prefix_conf.rds}, (ii) the gene 
mapping object config \code{prefix_gene.rds}, (iii) the single-cell gene 
expression \code{prefix_gexpr.h5}, (iv) the single-cell metadata 
\code{prefix_meta.rds} and (v) the defaults for the Shiny app 
\code{prefix_def.rds}. A prefix is specified in each file to allow for the 
use of multiple single-cell datasets in one Shiny app. Note that both 
\code{makeShinyFiles} and \code{makeShinyCodes} functions are ran when 
running the wrapper function \code{makeShinyApp}.
}
\examples{
makeShinyFiles(seu, scConf, gex.assay = "RNA", gex.slot = "data",
               shiny.prefix = "sc1", shiny.dir = "shinyApp/",
               default.gene1 = "GATA3", default.gene2 = "DNMT3L",
               default.multigene = c("ANPEP","NANOG","ZIC2","NLGN4X","DNMT3L",
                                     "DPPA5","SLC7A2","GATA3","KRT19"),
               default.dimred = c("UMAP_1", "UMAP_2"))

}
\author{
John F. Ouyang
}
