% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyCellPlus-package.R
\docType{package}
\name{ShinyCellPlus-package}
\alias{ShinyCellPlus}
\alias{ShinyCellPlus-package}
\title{ShinyCellPlus: Updated Shiny Web Apps for Single-Cell and Multiomic Data}
\description{
Shiny apps for interactive exploration of single-cell data. Based on the ShinyCell Project by John F. Ouyang \url{https://github.com/SGDDNB/ShinyCell}
}
\keyword{internal}
