% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrSVdeGenes}
\alias{wrSVdeGenes}
\title{Insert data server cues for differential expression genes table tab}
\usage{
wrSVdeGenes(prefix, de.genes)
}
\arguments{
\item{prefix}{file prefix}

\item{de.genes}{TRUE/FALSE whether to create this tab}
}
\description{
Insert data server cues for differential expression genes table tab
}
