% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrSVgeneOnt}
\alias{wrSVgeneOnt}
\title{Insert data server cues for ToppGene ontology tab}
\usage{
wrSVgeneOnt(prefix, gene.ont)
}
\arguments{
\item{prefix}{file prefix}

\item{gene.ont}{TRUE/FALSE whether to create this tab}
}
\description{
Insert data server cues for ToppGene ontology tab
}
