% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrSVmarkersAll}
\alias{wrSVmarkersAll}
\title{Insert data server cues for seurat cluster markers, all}
\usage{
wrSVmarkersAll(prefix, markers.all)
}
\arguments{
\item{prefix}{file prefix}

\item{markers.all}{TRUE/FALSE whether to create this tab}
}
\description{
Insert data server cues for seurat cluster markers, all
}
