% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrSVmarkersDE}
\alias{wrSVmarkersDE}
\title{Insert data server cues for marker and DE genes data tables}
\usage{
wrSVmarkersDE(prefix, markers.all, markers.top20, de.genes)
}
\description{
Insert data server cues for marker and DE genes data tables
}
