% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrSVmarkersTop20}
\alias{wrSVmarkersTop20}
\title{Insert data server cues for seurat cluster markers, top 20}
\usage{
wrSVmarkersTop20(prefix, markers.top20)
}
\arguments{
\item{prefix}{file prefix}

\item{markers.top20}{TRUE/FALSE whether to create this tab}
}
\description{
Insert data server cues for seurat cluster markers, top 20
}
