% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrSVvolc}
\alias{wrSVvolc}
\title{Insert data server cues for differential gene expression volcano plot tab}
\usage{
wrSVvolc(prefix, volc.plot)
}
\arguments{
\item{prefix}{file prefix}

\item{volc.plot}{TRUE/FALSE whether to create this tab}
}
\description{
Insert data server cues for differential gene expression volcano plot tab
}
