% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrUIdeGenes}
\alias{wrUIdeGenes}
\title{Append differential gene expression to main block of ui.R}
\usage{
wrUIdeGenes(prefix, de.genes)
}
\arguments{
\item{prefix}{file prefix}

\item{markers.top20}{TRUE/FALSE whether to include this tab}
}
\description{
Append differential gene expression to main block of ui.R
}
