% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrUIgeneOnt}
\alias{wrUIgeneOnt}
\title{Append ToppGene ontology page to ui.R}
\usage{
wrUIgeneOnt(prefix, gene.ont)
}
\arguments{
\item{prefix}{file prefix}

\item{gene.ont}{TRUE/FALSE whether to include this tab}
}
\description{
Append ToppGene ontology page to ui.R
}
