% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrUImarkersDE}
\alias{wrUImarkersDE}
\title{Append markers and DE tabs to main block of ui.R}
\usage{
wrUImarkersDE(prefix, markers.all, markers.top20, de.genes)
}
\description{
Append markers and DE tabs to main block of ui.R
}
