% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrUImarkersTop20}
\alias{wrUImarkersTop20}
\title{Append markers, top 20 to main block of ui.R}
\usage{
wrUImarkersTop20(prefix, markers.top20)
}
\arguments{
\item{prefix}{file prefix}

\item{markers.top20}{TRUE/FALSE whether to include this tab}
}
\description{
Append markers, top 20 to main block of ui.R
}
