% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrUIvolc}
\alias{wrUIvolc}
\title{Append gene differential gene expression volcano plot page to ui.R}
\usage{
wrUIvolc(prefix, volc.plot)
}
\arguments{
\item{prefix}{file prefix}

\item{gene.ranks}{TRUE/FALSE whether to include this tab}
}
\description{
Append gene differential gene expression volcano plot page to ui.R
}
