% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrSVgeneSig}
\alias{wrSVgeneSig}
\title{Insert data server cues for gene signature manipulate via AUCell}
\usage{
wrSVgeneSig(prefix, gene.ranks)
}
\description{
Insert data server cues for gene signature manipulate via AUCell
}
