% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrUImarkersAll}
\alias{wrUImarkersAll}
\title{Append markers, all to the main block of ui.R}
\usage{
wrUImarkersAll(prefix, markers.all)
}
\arguments{
\item{prefix}{file prefix}

\item{markers.all}{TRUE/FALSE whether to include this tab}
}
\description{
Append markers, all to the main block of ui.R
}
